Refactor : request 패키지 생성후 해당 클래스들 이동
