+++ src/main/java/controller/UserController.java
@@ -2,6 +2,10 @@ package controller;
 
 import db.Database;
 import http.*;
+import http.request.HttpRequest;
+import http.request.HttpUri;
+import http.request.QueryParameters;
+import http.request.RequestLine;
 import model.User;
 
 import java.util.HashMap;
