+++ src/main/java/http/request/HttpRequest.java
@@ -1,8 +1,7 @@
-package http;
+package http.request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.RequestHandler;
 
 import java.io.BufferedReader;
 import java.io.IOException;
