+++ src/main/java/http/request/HttpUri.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 public class HttpUri {
