+++ src/main/java/http/request/RequestLine.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 public class RequestLine {
