+++ src/main/java/http/request/ResourceType.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import exception.ResourceTypeNotFoundException;
