+++ src/test/java/HttpRequestTest.java
@@ -1,4 +1,4 @@
-import http.*;
+import http.request.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
